moved test cases all under one test file
remove unnecessary test cases, unify setup procedures
